Fix type error
Review URL: https://chromiumcodereview.appspot.com//9600057

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@5050 260f80e4-7a28-3924-810f-c04153c831b5

+++ dart/frog/leg/ssa/nodes.dart
@@ -173,7 +173,7 @@ class HGraph {
     return addConstant(new IntConstant(i));
   }
 
-  HConstant addConstantDouble(int d) {
+  HConstant addConstantDouble(double d) {
     return addConstant(new DoubleConstant(d));
   }
